refactor: decouple editor logic into modular components and hooks for better maintainability
